# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OCPCTL is a self-service OpenShift cluster provisioning and management platform. This API provides endpoints for creating, managing, and destroying OpenShift clusters with automated lifecycle management, policy enforcement, and resource tracking.",
        "title": "OCPCTL API",
        "termsOfService": "http://github.com/tsanders-rh/ocpctl",
        "contact": {
            "name": "OCPCTL Support",
            "url": "http://github.com/tsanders-rh/ocpctl/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/profiles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all available cluster profiles. Can be filtered by platform (aws or ibmcloud).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "List cluster profiles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by platform (aws, ibmcloud)",
                        "name": "platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.ProfileResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid platform parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/profiles/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns details of a specific cluster profile by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Get cluster profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ProfileResponse"
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_tsanders-rh_ocpctl_internal_profile.BaseDomainConfig": {
            "type": "object",
            "required": [
                "allowed",
                "default"
            ],
            "properties": {
                "allowed": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "default": {
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.ClusterNetworkConfig": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "string"
                },
                "hostPrefix": {
                    "type": "integer"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.ComputeConfig": {
            "type": "object",
            "required": [
                "control_plane",
                "workers"
            ],
            "properties": {
                "control_plane": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.ControlPlaneConfig"
                },
                "workers": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.WorkersConfig"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.ControlPlaneConfig": {
            "type": "object",
            "required": [
                "instance_type",
                "replicas"
            ],
            "properties": {
                "instance_type": {
                    "type": "string"
                },
                "replicas": {
                    "type": "integer",
                    "minimum": 1
                },
                "schedulable": {
                    "type": "boolean"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.CostControlsConfig": {
            "type": "object",
            "properties": {
                "budget_alert_threshold": {
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "estimated_hourly_cost": {
                    "type": "number"
                },
                "max_monthly_cost": {
                    "type": "number"
                },
                "warning_message": {
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.CustomResourceConfig": {
            "type": "object",
            "required": [
                "api_version",
                "kind",
                "name"
            ],
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "spec": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.FeaturesConfig": {
            "type": "object",
            "properties": {
                "fips_mode": {
                    "type": "boolean"
                },
                "off_hours_scaling": {
                    "type": "boolean"
                },
                "private_cluster": {
                    "type": "boolean"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.HelmChartConfig": {
            "type": "object",
            "required": [
                "chart",
                "name",
                "repo"
            ],
            "properties": {
                "chart": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                },
                "values": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.LifecycleConfig": {
            "type": "object",
            "required": [
                "default_ttl_hours",
                "max_ttl_hours"
            ],
            "properties": {
                "allow_custom_ttl": {
                    "type": "boolean"
                },
                "default_ttl_hours": {
                    "type": "integer",
                    "minimum": 1
                },
                "max_ttl_hours": {
                    "type": "integer",
                    "minimum": 1
                },
                "warn_before_destroy_hours": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.MachineNetworkConfig": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.ManifestConfig": {
            "type": "object",
            "required": [
                "name",
                "path"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.NetworkingConfig": {
            "type": "object",
            "properties": {
                "clusterNetworks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.ClusterNetworkConfig"
                    }
                },
                "machineNetwork": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.MachineNetworkConfig"
                    }
                },
                "networkType": {
                    "type": "string"
                },
                "serviceNetwork": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.OperatorConfig": {
            "type": "object",
            "required": [
                "channel",
                "name",
                "namespace",
                "source"
            ],
            "properties": {
                "channel": {
                    "type": "string"
                },
                "custom_resource": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.CustomResourceConfig"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "source": {
                    "description": "e.g. \"redhat-operators\"",
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.PostDeploymentConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "helm_charts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.HelmChartConfig"
                    }
                },
                "manifests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.ManifestConfig"
                    }
                },
                "operators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.OperatorConfig"
                    }
                },
                "timeout": {
                    "description": "Duration string, e.g. \"30m\"",
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.RegionConfig": {
            "type": "object",
            "required": [
                "allowed",
                "default"
            ],
            "properties": {
                "allowed": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "default": {
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.TagsConfig": {
            "type": "object",
            "properties": {
                "allow_user_tags": {
                    "type": "boolean"
                },
                "defaults": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "required": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.VersionConfig": {
            "type": "object",
            "required": [
                "allowed",
                "default"
            ],
            "properties": {
                "allowed": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "default": {
                    "type": "string"
                }
            }
        },
        "github_com_tsanders-rh_ocpctl_internal_profile.WorkersConfig": {
            "type": "object",
            "required": [
                "instance_type"
            ],
            "properties": {
                "autoscaling": {
                    "type": "boolean"
                },
                "instance_type": {
                    "type": "string"
                },
                "max_replicas": {
                    "type": "integer"
                },
                "min_replicas": {
                    "type": "integer",
                    "minimum": 0
                },
                "replicas": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "internal_api.ProfileResponse": {
            "type": "object",
            "properties": {
                "base_domains": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.BaseDomainConfig"
                },
                "compute": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.ComputeConfig"
                },
                "cost_controls": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.CostControlsConfig"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "features": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.FeaturesConfig"
                },
                "lifecycle": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.LifecycleConfig"
                },
                "name": {
                    "type": "string"
                },
                "networking": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.NetworkingConfig"
                },
                "openshift_versions": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.VersionConfig"
                },
                "platform": {
                    "type": "string"
                },
                "post_deployment": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.PostDeploymentConfig"
                },
                "regions": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.RegionConfig"
                },
                "tags": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_internal_profile.TagsConfig"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Format: \"Bearer {token}\". Obtain from /api/v1/auth/login endpoint.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}